+++ src/PatpassComercio/Inscription.php
@@ -17,6 +17,7 @@ use Transbank\Utils\HttpClientRequestService;
 use Transbank\Utils\RequestServiceTrait;
 use Transbank\Contracts\RequestService;
 use Transbank\PatpassComercio\Options;
+use Transbank\Utils\Curl\Exceptions\CurlRequestException;
 
 class Inscription
 {
@@ -62,7 +63,7 @@ class Inscription
      * @param string $city
      *
      * @throws InscriptionStartException
-     * @throws \Transbank\Utils\Curl\Exceptions\CurlRequestException
+     * @throws CurlRequestException
      *
      * @return InscriptionStartResponse
      */
@@ -122,7 +123,7 @@ class Inscription
      * @param string $token
      *
      * @throws InscriptionStatusException
-     * @throws \Transbank\Utils\Curl\Exceptions\CurlRequestException
+     * @throws CurlRequestException
      *
      * @return InscriptionStatusResponse
      */
