+++ src/Utils/Curl/Request.php
@@ -69,4 +69,24 @@ class Request implements RequestInterface
 
         return $new;
     }
+
+    public function withHeader($name, $value): RequestInterface
+    {
+        return clone $this;
+    }
+
+    public function withAddedHeader($name, $value): RequestInterface
+    {
+        return clone $this;
+    }
+
+    private function createBody($body = ''): StreamInterface
+    {
+        $resource = fopen('php://temp', 'rw+');
+        if (!empty($body)) {
+            fwrite($resource, $body);
+        }
+
+        return new Stream($resource);
+    }
 }
