+++ src/Utils/Curl/Response.php
@@ -46,4 +46,28 @@ class Response implements ResponseInterface
     {
         return $this->reasonPhrase;
     }
+
+    public function withHeader($name, $value): ResponseInterface
+    {
+        $new = clone $this;
+        $new->headers[$name] = (array) $value;
+        return $new;
+    }
+
+    public function withAddedHeader($name, $value): ResponseInterface
+    {
+        $new = clone $this;
+        $new->headers[$name][] = $value;
+        return $new;
+    }
+
+    private function createBody($body = ''): StreamInterface
+    {
+        $resource = fopen('php://temp', 'rw+');
+        if (!empty($body)) {
+            fwrite($resource, $body);
+        }
+
+        return new Stream($resource);
+    }
 }
