+++ src/Webpay/TransaccionCompleta/Transaction.php
@@ -16,6 +16,7 @@ use Transbank\Webpay\TransaccionCompleta\Responses\TransactionStatusResponse;
 use Transbank\Webpay\TransaccionCompleta\Responses\TransactionCaptureResponse;
 use Transbank\Utils\InteractsWithWebpayApi;
 use Transbank\Webpay\Exceptions\WebpayRequestException;
+use Transbank\Utils\Curl\Exceptions\CurlRequestException;
 
 /**
  * Class Transaction.
@@ -41,7 +42,7 @@ class Transaction
      * @param string|null   $cvv
      *
      * @throws TransactionCreateException
-     * @throws \Transbank\Utils\Curl\Exceptions\CurlRequestException
+     * @throws CurlRequestException
      *
      * @return TransactionCreateResponse
      */
@@ -82,7 +83,7 @@ class Transaction
      * @param int    $installmentsNumber
      *
      * @throws TransactionInstallmentsException
-     * @throws \Transbank\Utils\Curl\Exceptions\CurlRequestException
+     * @throws CurlRequestException
      *
      * @return TransactionInstallmentsResponse
      */
@@ -118,7 +119,7 @@ class Transaction
      * @param bool|null     $gracePeriod
      *
      * @throws TransactionCommitException
-     * @throws \Transbank\Utils\Curl\Exceptions\CurlRequestException
+     * @throws CurlRequestException
      *
      * @return TransactionCommitResponse
      */
@@ -156,7 +157,7 @@ class Transaction
      * @param int|float $amount
      *
      * @throws TransactionRefundException
-     * @throws \Transbank\Utils\Curl\Exceptions\CurlRequestException
+     * @throws CurlRequestException
      *
      * @return TransactionRefundResponse
      */
@@ -187,7 +188,7 @@ class Transaction
      * @param string $token
      *
      * @throws TransactionStatusException
-     * @throws \Transbank\Utils\Curl\Exceptions\CurlRequestException
+     * @throws CurlRequestException
      *
      * @return TransactionStatusResponse
      */
@@ -217,7 +218,7 @@ class Transaction
      * @param int|float $captureAmount
      *
      * @throws TransactionCaptureException
-     * @throws \Transbank\Utils\Curl\Exceptions\CurlRequestException
+     * @throws CurlRequestException
      *
      * @return TransactionCaptureResponse
      */
