+++ tests/Utils/Curl/RequestTest.php
@@ -3,6 +3,7 @@
 use PHPUnit\Framework\TestCase;
 use Transbank\Utils\Curl\Request;
 use Transbank\Utils\Curl\Uri;
+use Transbank\Utils\Curl\Stream;
 
 
 class RequestTest extends TestCase
@@ -59,12 +60,10 @@ class RequestTest extends TestCase
         $this->assertEquals([], $newRequest->getHeader('Accept'));
         $this->assertNotSame($newRequest, $this->request);
 
-        $secondRequest =
-            new Request('GET', 'https://www.transbank.cl:443/webpay/1.2/transactions/token?param1=123&param2=222', [
-                'Accept' => 'text/plain',
-                'api_key' => 'fakeApiKey'
-            ], 'this is a new body', '1.2');
-        $newRequest = $this->request->withBody($secondRequest->getBody());
+
+        $resource = fopen('php://temp', 'rw+');
+        fwrite($resource, 'testData');
+        $newRequest = $this->request->withBody(new Stream($resource));
         $this->assertFalse($newRequest->getBody() == $this->request->getBody());
         $this->assertNotSame($newRequest, $this->request);
     }
